💥  Lots of fun stylig
I love tailwind, it makes me start liking CSS.

I literally did CSS first this time because of how good it is.

I still need to implement all the JS lol and I'm procrastinating on it with CSS.

Yes. That's how good tailwind is.

I regret not trying this earlier.